[scripts] Have `test` paths match execution paths
The build script tests that developer tools exist at certain paths by
using `test "$FOO"`--note the quotes around `$FOO`. When invoking those
tools, on the other hand, the paths to the tools are *not* wrapped in
quotes. This causes script failures when the paths include spaces.

Wrap the tool paths in quotes, to match how they are tested.